Inital commit
This library is to allow easy rumbling of Xbox Controllers. The easiest
use case is that of notifying drivers of different events such as
value changes, warnings, etc. This library automatically rumbles with
the set intensity on the set side, then turns it off quickly to simulate
the rumbling in regular video games. Also, it can be run asynchronously
to not impede with the rest of the regular robot operations.