Add additional methods to Drivetrain
Add setNeutralMode and getNeutralMode methods, which sets and gets the
neutral mode of all the motors. The constructor now automatically sets
the neutral mode to coast, so the calls in RobotMap are removed.

The Drivetrain.constrainAngle function was also added, which constrains
an angle to the range (-180, 180]. getHeading now calls this method.