Fix bugs in RotateToAngle
The RotateToAngle command used to start an instance of FollowTrajectory.
But since FollowTrajectory also requires drivetrain, RotateToAngle would
be interrupted. Although the outcome is the same, RotateToAngle has been
changed to simply call methods from FollowTrajectory without starting
the command so that if future implementations want to add things in
isFinished(), interrupted() or end() there won't be an unexplained bug.