Change wait for vision initialization behavior
Instead of relying on wait() and notify() for vision subsystem
initialization, the new code will constantly check the status and will
skip the wait if the start button on the operator controller is pressed.
This is to fix a possible bug with the wait() and notify() method
causing the robot to be unresponsive.