Fix various bugs
Fix bug in FollowTrajectory where the command will not do anything after
the first time when the target is not found; fix bug in EssieDefault
where Essie's motors do not get stopped when the analog stick is back to
neutral.

Rewrite most of FollowTrajectory in an attempt to eliminate bugs; the
concurrently generated trajectories are now generated within the command
execution loop itself. This might be changed in the future.