Fix bug in TeleopDrive and change brake/coast behavior
Fix bug in TeleopDrive where the forwards/backwards control had an
exponent of 4 and left/right control was squared, instead of the other
way around.

Change Brake/Coast behavior of drivetrain motors. All auto commands that
involve driving will set the robot to be in brake, and the robot will
remain that way until in TeleopDrive a joystick is pushed past the
deadzone. This allows for more time to brake after an auto command ends
without making teleop driving brake as well.