Modify ConditionalButton's constructor
Modify ConditionalButton's three-argument constructor so that it's less
confusing. The previous constructor takes a boolean as the last
parameter, and if set to true, the ConditionalButton will only activate
when the condition is false. Even though it was documented, this
behavior is confusing. Update ConditionalButton's constructor so the
last parameter means what the boolean's state should be for the button
to activate instead of whether to invert.