Resolve issue #58
Resolve issue #58, "Gear state not updated in guest mode brach". This
issue was caused by the drivetrain gear entry only being updated when in
debug mode, which is a mistake. Although this is only known to happen on
the guest-mode branch, analysis of the code shows that it is an issue on
all branches.